general cleanup, d_menu_quit / d_a_obj_testcube mostly done, d_msg_scrn_explain debug (#3065)
* typedef for cPhs_Step

* make sdk includes consistent

* d_menu_quit / d_msg_scrn_explain debug

* d_a_obj_testcube mostly done

* d_debug_pad mostly done

* jstudio tool library headers

* some JStudioCameraEditor headers

* d_jcam_editor mostly done

* try fixing some shield regressions

* d_bg_parts mostly done

* fix merge errors

* debug fix

+++ include/JSystem/J3DGraphBase/J3DShape.h
@@ -189,7 +189,7 @@ public:
     J3DMaterial* getMaterial() const { return mMaterial; }
     u16 getIndex() const { return mIndex; }
     u32 getTexMtxLoadType() const { return mFlags & 0xF000; }
-    u32 getMtxGroupNum() const { return mMtxGroupNum; }
+    u16 getMtxGroupNum() const { return mMtxGroupNum; }
     J3DShapeDraw* getShapeDraw(u16 idx) { return mShapeDraw[idx]; }
     J3DShapeMtx* getShapeMtx(u16 idx) { return mShapeMtx[idx]; }
     Vec* getMin() { return &mMin; }
